+++ labbench/testing/pyvisa_sim.py
@@ -19,7 +19,7 @@ __all__ = ["PowerSensor", "Oscilloscope", "SignalGenerator"]
 # set the automatic connection filters
 @attr.adjust("make", "FakeTech")
 @attr.adjust("model", "Power Sensor #1234")
-@attr.adjust("write_termination", '\r\n')
+@attr.adjust("write_termination", "\r\n")
 class PowerSensor(VISADevice):
     RATES = "NORM", "DOUB", "FAST"
 
@@ -27,19 +27,23 @@ class PowerSensor(VISADevice):
     # taken from the instrument programming manual
     initiate_continuous = attr.property.bool(key="INIT:CONT", help="trigger continuously if True")
     trigger_count = attr.property.int(
-        key="TRIG:COUN", help="acquisition count", label="samples",
-        min=1, max=200
+        key="TRIG:COUN", help="acquisition count", label="samples", min=1, max=200
     )
     measurement_rate = attr.property.str(
-        key="SENS:MRAT", only=RATES, case=False,
+        key="SENS:MRAT",
+        only=RATES,
+        case=False,
     )
     sweep_aperture = attr.property.float(
-        key="SWE:APER", help="measurement duration", label="s",
-        min=20e-6, max=200e-3
+        key="SWE:APER", help="measurement duration", label="s", min=20e-6, max=200e-3
     )
     frequency = attr.property.float(
-        key="SENS:FREQ", help="calibration frequency", label="Hz",
-        min=10e6, max=18e9, step=1e-3,
+        key="SENS:FREQ",
+        help="calibration frequency",
+        label="Hz",
+        min=10e6,
+        max=18e9,
+        step=1e-3,
     )
 
     def preset(self):
@@ -97,7 +101,7 @@ class SpectrumAnalyzer(VISADevice):
         return series
 
     def trigger(self):
-        util.sleep(0.1) # simulate slow response
+        util.sleep(0.1)  # simulate slow response
         return self.write("TRIG")
 
 
@@ -107,8 +111,12 @@ class SpectrumAnalyzer(VISADevice):
 class SignalGenerator(VISADevice):
     output_enabled = attr.property.bool(key="OUT:ENABL", help="when True, output an RF tone")
     center_frequency = attr.property.float(
-        key="SENS:FREQ", help="input signal center frequency", label="Hz",
-        min=10e6, max=18e9, step=1e-3,
+        key="SENS:FREQ",
+        help="input signal center frequency",
+        label="Hz",
+        min=10e6,
+        max=18e9,
+        step=1e-3,
     )
     mode = attr.property.str(key="MODE", only=["sweep", "tone", "iq"], case=False)
 
@@ -123,8 +131,11 @@ class SignalGenerator(VISADevice):
 @attr.adjust("model", default="Oscilloscope #1234")
 class Oscilloscope(VISADevice):
     @attr.method.float(
-        label="Hz", help="channel center frequency",
-        min=10e6, max=18e9, step=1e-3,                       
+        label="Hz",
+        help="channel center frequency",
+        min=10e6,
+        max=18e9,
+        step=1e-3,
     )
     def center_frequency(self, set_value=Undefined, /, *, channel):
         if set_value is Undefined:
@@ -133,8 +144,12 @@ class Oscilloscope(VISADevice):
             self.write(f"CH{channel}:SENS:FREQ {set_value}")
 
     resolution_bandwidth = attr.method.float(
-        key="CH{channel}:SENS:BW", help="channel resolution bandwidth", label="Hz",
-        min=1, max=40e6, step=1e-3
+        key="CH{channel}:SENS:BW",
+        help="channel resolution bandwidth",
+        label="Hz",
+        min=1,
+        max=40e6,
+        step=1e-3
         # arguments omitted deliberately for testing
     )